Moved components and small refactoring
