+++ src/components/SkillItem/SkillItem.scss
@@ -1,4 +1,4 @@
-@import "public/scss/variables";
+@import "../../../public/scss/variables";
 
 .skill-item {padding: 10px 20px; border-radius: 100px; border: 1px solid $primary;
   transition: all .3s; display: flex; vertical-align: middle; margin: 5px;
