+++ src/containers/SkillItems/SkillItems.scss
@@ -1,4 +1,4 @@
-@import "public/scss/variables";
+@import "../../../public/scss/variables";
 
 .skill-items {
   @media (max-width: $mobile) {padding-top: 50px;}
